Close to last revision

+++ BubbleShot/src/mainGame/frontend/fxmls/PlayMenuController.java
@@ -4,8 +4,11 @@ import java.net.URL;
 import java.util.Date;
 import java.util.ResourceBundle;
 
+import javafx.animation.KeyFrame;
+import javafx.animation.Timeline;
 import javafx.fxml.Initializable;
 import javafx.scene.Group;
+import javafx.scene.control.Label;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.Skin;
 import javafx.scene.image.Image;
@@ -14,6 +17,7 @@ import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
 import javafx.scene.text.Text;
+import javafx.util.Duration;
 import mainGame.GameRunner;
 import mainGame.SceneTracker;
 import mainGame.saving.FileDeleter;
@@ -27,11 +31,16 @@ public class PlayMenuController implements Initializable
 	public BorderPane container;
 	public ScrollPane scrollPaneContainer;
 	
+	public Label messageLabel;
 	public VBox playerPickerRow;
 	private ImageView[] characterImages;
 	private HBox[] characterSaveSlots;
 	private Player[] players;
 	private Player selectedPlayer;
+	private Timeline timeline;
+	private KeyFrame kf;
+	
+	private int deleteBtnClick;
 	
 	@Override
 	public void initialize(URL location, ResourceBundle resources)
@@ -40,9 +49,9 @@ public class PlayMenuController implements Initializable
 		container.setPrefWidth(GameRunner.getResolutionWidth());
 		container.setId("borderPaneContainer");
 		scrollPaneContainer.prefWidthProperty().bind(container.prefWidthProperty());
-		
+		deleteBtnClick = 0;
 		players = FileReader.loadPlayer();
-		if(players != null)
+		if(players != null && players.length > 0)
 		{
 			characterImages = new ImageView[players.length];
 			characterSaveSlots = new HBox[players.length];
@@ -50,7 +59,7 @@ public class PlayMenuController implements Initializable
 			{
 				Text playerStatTxt = new Text("Name " + players[i].getSpriteName()+ "\t Current Level " + 
 						players[i].getLocalLevel() + "\t Coins " + players[i].getCoins());
-				Text dateTxt = new Text("Last Played: " + players[i].getDate().toString());
+				Text dateTxt = new Text("Last Saved: " + players[i].getDate().toString());
 				characterImages[i] = new ImageView(players[i].getSpriteImage());
 				characterImages[i].setFitWidth(100);
 				characterImages[i].setFitHeight(100);
@@ -68,7 +77,9 @@ public class PlayMenuController implements Initializable
 						characterSaveSlots[j].setStyle("-fx-border-width:5px");
 						if(characterSaveSlots[j].equals(event.getSource()))
 						{
+							SoundEffects.MOUSE_CLICK_SOUND.playSound();
 							selectedPlayer = players[j];
+							deleteBtnClick = 0;
 							characterSaveSlots[j].setStyle("-fx-border-width:10px");
 						}
 					}
@@ -81,6 +92,22 @@ public class PlayMenuController implements Initializable
 				playerPickerRow.getChildren().add(characterSaveSlots[i]);
 			}
 		}
+		else
+		{
+			messageLabel.setText("No Characters");
+		}
+		generateClearMessageLabelTimer();
+	}
+	
+	private void generateClearMessageLabelTimer()
+	{
+		timeline = new Timeline();
+		kf = new KeyFrame(Duration.seconds(2), event -> 
+		{
+			messageLabel.setText("");
+			deleteBtnClick = 0;
+		});
+		timeline.getKeyFrames().add(kf);
 	}
 	
 	private void displaySelectedPlayer()
@@ -109,7 +136,11 @@ public class PlayMenuController implements Initializable
 	public void startBtnOnclick()
 	{
 		if(selectedPlayer == null)
+		{
+			messageLabel.setText("Select A Character Before Starting");
+			clearMessage();
 			return;
+		}
 		SoundEffects.MENU_SELECT_SOUND.playSound();
 		Player.setCurrentLevel(selectedPlayer.getLocalLevel());
 		GameRunner.createGameManager(LevelDesign.getRandomLevelDesign(4, 4, selectedPlayer.getLocalLevel()), selectedPlayer);
@@ -121,10 +152,28 @@ public class PlayMenuController implements Initializable
 	public void deleteBtnOnClick() throws Exception
 	{
 		if(selectedPlayer == null)
+		{
+			messageLabel.setText("Select A Character Before Deleting");
+			clearMessage();
 			return;
-		SoundEffects.MENU_SELECT_SOUND.playSound();
-		FileDeleter.deletePlayer(selectedPlayer);
-		GameRunner.getSceneTracker().switchToPlayMenuView();
+		}
+		if(deleteBtnClick == 1)
+		{
+			SoundEffects.MENU_SELECT_SOUND.playSound();
+			FileDeleter.deletePlayer(selectedPlayer);
+			GameRunner.getSceneTracker().switchToPlayMenuView();
+		}
+		else
+		{
+			deleteBtnClick = 1;
+			messageLabel.setText("Press Again To Delete");
+			clearMessage();
+		}
+	}
+	
+	public void clearMessage()
+	{
+		timeline.playFromStart();
 	}
 	
 }
